+++ packages/eslint-plugin-isaacscript/src/configs/recommended.ts
@@ -22,7 +22,7 @@ export const recommended = {
     "isaacscript/no-explicit-map-set-loops": "error",
     "isaacscript/no-for-in": "error",
     "isaacscript/no-let-any": "error",
-    "isaacscript/no-mutable-array-return": "error",
+    "isaacscript/no-mutable-return": "error",
     "isaacscript/no-number-enums": "error",
     "isaacscript/no-object-any": "error",
     "isaacscript/no-object-methods-with-map-set": "error",