+++ packages/isaacscript-cli/src/commands/monitor/saveDatWriter/saveDatWriter.ts
@@ -60,7 +60,7 @@ function onMessage(type: SaveDatMessageType, data: string, numRetries = 0) {
   writeSaveDatToDisk(type, data, saveDat, numRetries);
 }
 
-// eslint-disable-next-line isaacscript/no-mutable-array-return
+// eslint-disable-next-line isaacscript/no-mutable-return
 function readSaveDatFromDisk(): SaveDatMessage[] {
   let saveDat: SaveDatMessage[];
   if (isFile(saveDatPath)) {