+++ packages/isaacscript-common/src/functions/set.ts
@@ -118,7 +118,7 @@ export function getSetCombinations<T>(
  * Normally, set values are returned in insertion order, so use this function when the ordering of
  * the contents is important.
  */
-// eslint-disable-next-line isaacscript/no-mutable-array-return
+// eslint-disable-next-line isaacscript/no-mutable-return
 export function getSortedSetValues<T>(set: ReadonlySet<T>): T[] {
   const values = [...set];